refactor!: replace riot.observable with Emitter
BREAKING CHANGE: this affects user code, specifically all event listeners that register exactly one event.
The positional args for these listeners need to be shifted to the right by one.

This change standardizes the event name as always being the first callback argument, instead of only
when multiple events are registered in one on() or once() call.